use webpack-log

+++ src/index.ts
@@ -1,45 +1,69 @@
+import ts from "typescript";
 import * as webpack from "webpack";
 import * as path from "path";
-import ts from "typescript";
 import * as event from "events";
-// https://nodejs.org/api/worker_threads.html
-const workerThreads = require("worker_threads");
+import * as workerThreads from "worker_threads"; // https://nodejs.org/api/worker_threads.html
+import { findTsConfig } from "./find_ts_config";
+import { logger } from "./logger";
 
 const NAME = "TsCheckerPlugin";
 export default class TsCheckerPlugin {
   exited: boolean = false;
   errors: any[] = [];
-  worker: any = null;
+  worker: workerThreads.Worker | null = null;
   watch: boolean = false;
   event: event.EventEmitter = new event.EventEmitter();
+  tsConfigPath?: string;
+  tsCompilerOptions?: ts.CompilerOptions;
 
-  constructor(options: { tsConfigPath: string; options: object }) {
+  constructor(options: { tsConfigPath?: string; options?: ts.CompilerOptions } = {}) {
+    this.tsConfigPath = options.tsConfigPath;
+    this.tsCompilerOptions = options.options;
     this.setupEventHandler();
   }
 
   apply(compiler: webpack.Compiler) {
     compiler.hooks.run.tapAsync(NAME, (compilation: webpack.compilation.Compilation, callback) => {
-      const tsConfigPath = this.resolveTsConfigPath(compiler);
-      // console.log(`tsConfigPath = ${tsConfigPath}`);
       this.watch = false;
       this.errors = [];
-      this.worker = this.hookWorkerEvents(this.hookWorkerCommonEvents(this.createWorker(tsConfigPath)));
-      callback();
+      this.resolveTsConfigPath(compiler)
+        .then(tsConfigPath => {
+          logger.info("tsConfigPath = ", tsConfigPath);
+          this.worker = this.hookWorkerEvents(this.hookWorkerCommonEvents(this.createWorker(tsConfigPath)));
+          callback();
+        })
+        .catch(error => {
+          logger.error(error);
+          callback();
+        });
     });
 
     compiler.hooks.watchRun.tapAsync(NAME, (compiler: webpack.Compiler, callback) => {
-      const tsConfigPath = this.resolveTsConfigPath(compiler);
-      // console.log(`tsConfigPath = ${tsConfigPath}`);
       this.watch = true;
       this.errors = [];
-      this.worker = this.hookWatchWorkerEvents(this.hookWorkerCommonEvents(this.createWatchWorker(tsConfigPath)));
-      callback();
+      if (this.worker) {
+        callback();
+        return;
+      }
+      this.resolveTsConfigPath(compiler)
+        .then(tsConfigPath => {
+          logger.info("tsConfigPath = ", tsConfigPath);
+          this.worker = this.hookWatchWorkerEvents(this.hookWorkerCommonEvents(this.createWatchWorker(tsConfigPath)));
+          callback();
+        })
+        .catch(error => {
+          logger.error(error);
+          callback();
+        });
     });
 
     compiler.hooks.watchClose.tap(NAME, () => {
-      const threadId = this.worker.threadId;
-      this.worker.unref();
-      this.worker.terminate(() => this.event.emit("worker:terminate", threadId));
+      logger.info("watchClose");
+      if (this.worker) {
+        const threadId = this.worker.threadId;
+        this.worker.unref();
+        this.worker.terminate(() => this.event.emit("worker:terminate", threadId));
+      }
       this.worker = null;
     });
 
@@ -53,10 +77,7 @@ export default class TsCheckerPlugin {
   }
 
   resolveTsConfigPath(compiler: webpack.Compiler) {
-    const tsConfigPath = compiler.options.context
-      ? path.resolve(compiler.options.context, "tsconfig.json")
-      : path.resolve("./tsconfig.json");
-    return tsConfigPath;
+    return findTsConfig([this.tsConfigPath, compiler.options.context, "."]);
   }
 
   waitForWatchDone(stats: webpack.Stats, callback) {
@@ -78,20 +99,15 @@ export default class TsCheckerPlugin {
     this.errors = [];
   }
 
-  unrefWorker() {
-    if (this.worker) {
-      this.worker.unref();
-      this.worker = null;
-    }
-  }
-
-  hookWorkerEvents(worker) {
+  hookWorkerEvents(worker: workerThreads.Worker) {
     worker.on("online", () => this.event.emit("start"));
     worker.on("exit", () => this.event.emit("done"));
     worker.on("error", () => this.event.emit("done"));
     worker.on("message", message => {
-      if (message.result === "diagnostics") {
-        message.diagnostics.forEach((diagnostic) => {
+      if (message.type === "log") {
+        logger.info("log", ...message.payload);
+      } else if (message.type === "diagnostics") {
+        message.payload.forEach(diagnostic => {
           this.errors.push(diagnostic);
         });
       }
@@ -99,32 +115,28 @@ export default class TsCheckerPlugin {
     return worker;
   }
 
-  hookWatchWorkerEvents(worker) {    
+  hookWatchWorkerEvents(worker) {
     worker.on("message", message => {
-      switch (message.result) {
-        case "report":
-          const diagnostic: ts.Diagnostic = message.diagnostic;
-          // console.log(diagnostic.code, diagnostic.messageText);
-          if (diagnostic.code === 6031) {
-            // Starting compilation in watch mode...
-            this.event.emit("start");
-          } else if (diagnostic.code === 6032) {
-            // File change detected. Starting incremental compilation...
-            this.event.emit("start");
-          } else if (diagnostic.code === 6193) { 
-            // Found 1 error. Watching for file changes.
-            this.event.emit("done");
-          } else if (diagnostic.code === 6194) {
-            // Found {0} errors. Watching for file changes.
-            this.event.emit("done");
-          } else {
-          }
-          break;
-        case "diagnostics":
-          message.diagnostics.forEach((diagnostic) => {
-            this.errors.push(diagnostic);
-          });
-          break;
+      if (message.type === "log") {
+        // logger.info("log", ...message.payload);
+      } else if (message.type === "diagnostic") {
+        this.errors.push(message.payload);
+      } else if (message.type === "report") {
+        const diagnostic: ts.Diagnostic = message.payload;
+        if (diagnostic.code === 6031) {
+          // Starting compilation in watch mode...
+          this.event.emit("start");
+        } else if (diagnostic.code === 6032) {
+          // File change detected. Starting incremental compilation...
+          this.event.emit("start");
+        } else if (diagnostic.code === 6193) {
+          // Found 1 error. Watching for file changes.
+          this.event.emit("done");
+        } else if (diagnostic.code === 6194) {
+          // Found {0} errors. Watching for file changes.
+          this.event.emit("done");
+        } else {
+        }
       }
     });
     return worker;
@@ -171,26 +183,31 @@ export default class TsCheckerPlugin {
   }
 
   setupEventHandler() {
+    const t: any = {};
     this.event.on("worker:new", threadId => {
-      (new Date());
+      t.new = Date.now();
+      logger.info("worker:new");
     });
     this.event.on("worker:online", threadId => {
-      (new Date());
+      t.online = Date.now();
+      logger.info("worker:online", `online - new = ${t.online - t.new}ms`);
     });
     this.event.on("worker:exit", ([threadId, exitCode]) => {
-      (new Date());
+      logger.info("worker:exit", "exitCode = ", exitCode);
     });
     this.event.on("worker:error", ([threadId, error]) => {
-      (new Date());
+      logger.info("worker:error", error);
     });
     this.event.on("worker:terminate", threadId => {
-      (new Date());
+      logger.info("worker:terminate");
     });
     this.event.on("start", () => {
-      (new Date());
+      t.start = Date.now();
+      logger.info("Start");
     });
     this.event.on("done", () => {
-      (new Date());
+      t.done = Date.now();
+      logger.info("Done", `${t.done - t.start}ms`);
     });
   }
 }
